pyproject.toml
[tool.poetry]
name = "inner_dialog"
version = "0.1.0"
description = ""
authors = ["8igmac <8igmac@pm.me>"]
readme = "README.md"
packages = [{include = "inner_dialog"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
openai = "^0.27.4"
langchain = "^0.0.175"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
google-generativeai = "^0.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.2.2"
tqdm = "^4.65.0"
bert-score = "^0.3.13"
sentence-transformers = "^2.2.2"
evaluate = "^0.4.0"
rouge-score = "^0.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
